+++ src/GameGUI/LootGUI.cpp
@@ -820,7 +820,7 @@ void LootGUI::LootItemReleased(LootSlotItem* pLootItem)
 	int y;
 	int width;
 	int height;
-	POINT lMouse = { VoxGame::GetInstance()->GetWindowCursorX(), (m_windowHeight-VoxGame::GetInstance()->GetWindowCursorY()) };
+	POINT lMouse = { (float)VoxGame::GetInstance()->GetWindowCursorX(), (float)(m_windowHeight-VoxGame::GetInstance()->GetWindowCursorY()) };
 	bool switched = false;
 	bool sameSlot = false;
 	bool draggedToInventory = false;
